requirements.txt
Django==4.0.5
django-filter==22.1
django-q==1.3.9
djangorestframework==3.13.1
Pillow==9.1.1
pytest==7.1.2
pytest-django==4.5.2